Struct rusoto_mturk::CreateHITRequest

pub struct CreateHITRequest {
    pub assignment_duration_in_seconds: i64,
    pub assignment_review_policy: Option<ReviewPolicy>,
    pub auto_approval_delay_in_seconds: Option<i64>,
    pub description: String,
    pub hit_layout_id: Option<String>,
    pub hit_layout_parameters: Option<Vec<HITLayoutParameter>>,
    pub hit_review_policy: Option<ReviewPolicy>,
    pub keywords: Option<String>,
    pub lifetime_in_seconds: i64,
    pub max_assignments: Option<i64>,
    pub qualification_requirements: Option<Vec<QualificationRequirement>>,
    pub question: Option<String>,
    pub requester_annotation: Option<String>,
    pub reward: String,
    pub title: String,
    pub unique_request_token: Option<String>,
}

Fields
assignment_duration_in_seconds: i64

The amount of time, in seconds, that a Worker has to complete the HIT after accepting it. If a Worker does not complete the assignment within the specified duration, the assignment is considered abandoned. If the HIT is still active (that is, its lifetime has not elapsed), the assignment becomes available for other users to find and accept.

assignment_review_policy: Option<ReviewPolicy>

The Assignment-level Review Policy applies to the assignments under the HIT. You can specify for Mechanical Turk to take various actions based on the policy.

auto_approval_delay_in_seconds: Option<i64>

The number of seconds after an assignment for the HIT has been submitted, after which the assignment is considered Approved automatically unless the Requester explicitly rejects it.

description: String

A general description of the HIT. A description includes detailed information about the kind of task the HIT contains. On the Amazon Mechanical Turk web site, the HIT description appears in the expanded view of search results, and in the HIT and assignment screens. A good description gives the user enough information to evaluate the HIT before accepting it.

hit_layout_id: Option<String>

The HITLayoutId allows you to use a pre-existing HIT design with placeholder values and create an additional HIT by providing those values as HITLayoutParameters.

Constraints: Either a Question parameter or a HITLayoutId parameter must be provided.

hit_layout_parameters: Option<Vec<HITLayoutParameter>>

If the HITLayoutId is provided, any placeholder values must be filled in with values using the HITLayoutParameter structure. For more information, see HITLayout.

hit_review_policy: Option<ReviewPolicy>

The HIT-level Review Policy applies to the HIT. You can specify for Mechanical Turk to take various actions based on the policy.

keywords: Option<String>

One or more words or phrases that describe the HIT, separated by commas. These words are used in searches to find HITs.

lifetime_in_seconds: i64

An amount of time, in seconds, after which the HIT is no longer available for users to accept. After the lifetime of the HIT elapses, the HIT no longer appears in HIT searches, even if not all of the assignments for the HIT have been accepted.

max_assignments: Option<i64>

The number of times the HIT can be accepted and completed before the HIT becomes unavailable.

qualification_requirements: Option<Vec<QualificationRequirement>>

A condition that a Worker's Qualifications must meet before the Worker is allowed to accept and complete the HIT.

question: Option<String>

The data the person completing the HIT uses to produce the results.

Constraints: Must be a QuestionForm data structure, an ExternalQuestion data structure, or an HTMLQuestion data structure. The XML question data must not be larger than 64 kilobytes (65,535 bytes) in size, including whitespace.

Either a Question parameter or a HITLayoutId parameter must be provided.

requester_annotation: Option<String>

An arbitrary data field. The RequesterAnnotation parameter lets your application attach arbitrary data to the HIT for tracking purposes. For example, this parameter could be an identifier internal to the Requester's application that corresponds with the HIT.

The RequesterAnnotation parameter for a HIT is only visible to the Requester who created the HIT. It is not shown to the Worker, or any other Requester.

The RequesterAnnotation parameter may be different for each HIT you submit. It does not affect how your HITs are grouped.

reward: String

The amount of money the Requester will pay a Worker for successfully completing the HIT.

title: String

The title of the HIT. A title should be short and descriptive about the kind of task the HIT contains. On the Amazon Mechanical Turk web site, the HIT title appears in search results, and everywhere the HIT is mentioned.

unique_request_token: Option<String>

A unique identifier for this request which allows you to retry the call on error without creating duplicate HITs. This is useful in cases such as network timeouts where it is unclear whether or not the call succeeded on the server. If the HIT already exists in the system from a previous call using the same UniqueRequestToken, subsequent calls will return a AWS.MechanicalTurk.HitAlreadyExists error with a message containing the HITId.

Note: It is your responsibility to ensure uniqueness of the token. The unique token expires after 24 hours. Subsequent calls using the same UniqueRequestToken made after the 24 hour limit could create duplicate HITs.