Struct rusoto_mturk::CreateQualificationTypeRequest

pub struct CreateQualificationTypeRequest {
    pub answer_key: Option<String>,
    pub auto_granted: Option<bool>,
    pub auto_granted_value: Option<i64>,
    pub description: String,
    pub keywords: Option<String>,
    pub name: String,
    pub qualification_type_status: String,
    pub retry_delay_in_seconds: Option<i64>,
    pub test: Option<String>,
    pub test_duration_in_seconds: Option<i64>,
}

Fields
answer_key: Option<String>

The answers to the Qualification test specified in the Test parameter, in the form of an AnswerKey data structure.

Constraints: Must not be longer than 65535 bytes.

Constraints: None. If not specified, you must process Qualification requests manually.

auto_granted: Option<bool>

Specifies whether requests for the Qualification type are granted immediately, without prompting the Worker with a Qualification test.

Constraints: If the Test parameter is specified, this parameter cannot be true.

auto_granted_value: Option<i64>

The Qualification value to use for automatically granted Qualifications. This parameter is used only if the AutoGranted parameter is true.

description: String

A long description for the Qualification type. On the Amazon Mechanical Turk website, the long description is displayed when a Worker examines a Qualification type.

keywords: Option<String>

One or more words or phrases that describe the Qualification type, separated by commas. The keywords of a type make the type easier to find during a search.

name: String

The name you give to the Qualification type. The type name is used to represent the Qualification to Workers, and to find the type using a Qualification type search. It must be unique across all of your Qualification types.

qualification_type_status: String

The initial status of the Qualification type.

Constraints: Valid values are: Active | Inactive

retry_delay_in_seconds: Option<i64>

The number of seconds that a Worker must wait after requesting a Qualification of the Qualification type before the worker can retry the Qualification request.

Constraints: None. If not specified, retries are disabled and Workers can request a Qualification of this type only once, even if the Worker has not been granted the Qualification. It is not possible to disable retries for a Qualification type after it has been created with retries enabled. If you want to disable retries, you must delete existing retry-enabled Qualification type and then create a new Qualification type with retries disabled.

test: Option<String>

The questions for the Qualification test a Worker must answer correctly to obtain a Qualification of this type. If this parameter is specified, TestDurationInSeconds must also be specified.

Constraints: Must not be longer than 65535 bytes. Must be a QuestionForm data structure. This parameter cannot be specified if AutoGranted is true.

Constraints: None. If not specified, the Worker may request the Qualification without answering any questions.

test_duration_in_seconds: Option<i64>

The number of seconds the Worker has to complete the Qualification test, starting from the time the Worker requests the Qualification.