Struct rusoto_mturk::HIT

pub struct HIT {
    pub assignment_duration_in_seconds: Option<i64>,
    pub auto_approval_delay_in_seconds: Option<i64>,
    pub creation_time: Option<f64>,
    pub description: Option<String>,
    pub expiration: Option<f64>,
    pub hit_group_id: Option<String>,
    pub hit_id: Option<String>,
    pub hit_layout_id: Option<String>,
    pub hit_review_status: Option<String>,
    pub hit_status: Option<String>,
    pub hit_type_id: Option<String>,
    pub keywords: Option<String>,
    pub max_assignments: Option<i64>,
    pub number_of_assignments_available: Option<i64>,
    pub number_of_assignments_completed: Option<i64>,
    pub number_of_assignments_pending: Option<i64>,
    pub qualification_requirements: Option<Vec<QualificationRequirement>>,
    pub question: Option<String>,
    pub requester_annotation: Option<String>,
    pub reward: Option<String>,
    pub title: Option<String>,
}

The HIT data structure represents a single HIT, including all the information necessary for a Worker to accept and complete the HIT.

Fields
assignment_duration_in_seconds: Option<i64>

The length of time, in seconds, that a Worker has to complete the HIT after accepting it.

auto_approval_delay_in_seconds: Option<i64>

The amount of time, in seconds, after the Worker submits an assignment for the HIT that the results are automatically approved by Amazon Mechanical Turk. This is the amount of time the Requester has to reject an assignment submitted by a Worker before the assignment is auto-approved and the Worker is paid.

creation_time: Option<f64>

The date and time the HIT was created.

description: Option<String>

A general description of the HIT.

expiration: Option<f64>

The date and time the HIT expires.

hit_group_id: Option<String>

The ID of the HIT Group of this HIT.

hit_id: Option<String>

A unique identifier for the HIT.

hit_layout_id: Option<String>

The ID of the HIT Layout of this HIT.

hit_review_status: Option<String>

Indicates the review status of the HIT. Valid Values are NotReviewed | MarkedForReview | ReviewedAppropriate | ReviewedInappropriate.

hit_status: Option<String>

The status of the HIT and its assignments. Valid Values are Assignable | Unassignable | Reviewable | Reviewing | Disposed.

hit_type_id: Option<String>

The ID of the HIT type of this HIT

keywords: Option<String>

One or more words or phrases that describe the HIT, separated by commas. Search terms similar to the keywords of a HIT are more likely to have the HIT in the search results.

max_assignments: Option<i64>

The number of times the HIT can be accepted and completed before the HIT becomes unavailable.

number_of_assignments_available: Option<i64>

The number of assignments for this HIT that are available for Workers to accept.

number_of_assignments_completed: Option<i64>

The number of assignments for this HIT that have been approved or rejected.

number_of_assignments_pending: Option<i64>

The number of assignments for this HIT that are being previewed or have been accepted by Workers, but have not yet been submitted, returned, or abandoned.

qualification_requirements: Option<Vec<QualificationRequirement>>

A condition that a Worker's Qualifications must meet in order to accept the HIT. A HIT can have between zero and ten Qualification requirements. All requirements must be met by a Worker's Qualifications for the Worker to accept the HIT.

question: Option<String>

The data the Worker completing the HIT uses produce the results. This is either either a QuestionForm, HTMLQuestion or an ExternalQuestion data structure.

requester_annotation: Option<String>

An arbitrary data field the Requester who created the HIT can use. This field is visible only to the creator of the HIT.

reward: Option<String> title: Option<String>

The title of the HIT.