Struct rusoto_mturk::MechanicalTurkClient

pub struct MechanicalTurkClient { /* fields omitted */ }

A client for the Amazon MTurk API.

Methods

impl MechanicalTurkClient

pub fn new(region: Region) -> MechanicalTurkClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> MechanicalTurkClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,