Struct rusoto_mturk::QualificationRequest

pub struct QualificationRequest {
    pub answer: Option<String>,
    pub qualification_request_id: Option<String>,
    pub qualification_type_id: Option<String>,
    pub submit_time: Option<f64>,
    pub test: Option<String>,
    pub worker_id: Option<String>,
}

The QualificationRequest data structure represents a request a Worker has made for a Qualification.

Fields
answer: Option<String>

The Worker's answers for the Qualification type's test contained in a QuestionFormAnswers document, if the type has a test and the Worker has submitted answers. If the Worker does not provide any answers, Answer may be empty.

qualification_request_id: Option<String>

The ID of the Qualification request, a unique identifier generated when the request was submitted.

qualification_type_id: Option<String>

The ID of the Qualification type the Worker is requesting, as returned by the CreateQualificationType operation.

submit_time: Option<f64>

The date and time the Qualification request had a status of Submitted. This is either the time the Worker submitted answers for a Qualification test, or the time the Worker requested the Qualification if the Qualification type does not have a test.

test: Option<String>

The contents of the Qualification test that was presented to the Worker, if the type has a test and the Worker has submitted answers. This value is identical to the QuestionForm associated with the Qualification type at the time the Worker requests the Qualification.

worker_id: Option<String>

The ID of the Worker requesting the Qualification.