Crate rusoto_mturk

Amazon Mechanical Turk API Reference

If you're using the service, you're probably looking for MechanicalTurkClient and MechanicalTurk.

Structs

AcceptQualificationRequestRequest

AcceptQualificationRequestResponse

ApproveAssignmentRequest

ApproveAssignmentResponse

Assignment
The Assignment data structure represents a single assignment of a HIT to a Worker. The assignment tracks the Worker's efforts to complete the HIT, and contains the results for later retrieval.

AssociateQualificationWithWorkerRequest

AssociateQualificationWithWorkerResponse

BonusPayment
An object representing a Bonus payment paid to a Worker.

CreateAdditionalAssignmentsForHITRequest

CreateAdditionalAssignmentsForHITResponse

CreateHITRequest

CreateHITResponse

CreateHITTypeRequest

CreateHITTypeResponse

CreateHITWithHITTypeRequest

CreateHITWithHITTypeResponse

CreateQualificationTypeRequest

CreateQualificationTypeResponse

CreateWorkerBlockRequest

CreateWorkerBlockResponse

DeleteHITRequest

DeleteHITResponse

DeleteQualificationTypeRequest

DeleteQualificationTypeResponse

DeleteWorkerBlockRequest

DeleteWorkerBlockResponse

DisassociateQualificationFromWorkerRequest

DisassociateQualificationFromWorkerResponse

GetAccountBalanceRequest

GetAccountBalanceResponse

GetAssignmentRequest

GetAssignmentResponse

GetFileUploadURLRequest

GetFileUploadURLResponse

GetHITRequest

GetHITResponse

GetQualificationScoreRequest

GetQualificationScoreResponse

GetQualificationTypeRequest

GetQualificationTypeResponse

HIT
The HIT data structure represents a single HIT, including all the information necessary for a Worker to accept and complete the HIT.

HITLayoutParameter
The HITLayoutParameter data structure defines parameter values used with a HITLayout. A HITLayout is a reusable Amazon Mechanical Turk project template used to provide Human Intelligence Task (HIT) question data for CreateHIT.

ListAssignmentsForHITRequest

ListAssignmentsForHITResponse

ListBonusPaymentsRequest

ListBonusPaymentsResponse

ListHITsForQualificationTypeRequest

ListHITsForQualificationTypeResponse

ListHITsRequest

ListHITsResponse

ListQualificationRequestsRequest

ListQualificationRequestsResponse

ListQualificationTypesRequest

ListQualificationTypesResponse

ListReviewPolicyResultsForHITRequest

ListReviewPolicyResultsForHITResponse

ListReviewableHITsRequest

ListReviewableHITsResponse

ListWorkerBlocksRequest

ListWorkerBlocksResponse

ListWorkersWithQualificationTypeRequest

ListWorkersWithQualificationTypeResponse

Locale
The Locale data structure represents a geographical region or location.

MechanicalTurkClient
A client for the Amazon MTurk API.

NotificationSpecification
The NotificationSpecification data structure describes a HIT event notification for a HIT type.

NotifyWorkersFailureStatus
When MTurk encounters an issue with notifying the Workers you specified, it returns back this object with failure details.

NotifyWorkersRequest

NotifyWorkersResponse

ParameterMapEntry
This data structure is the data type for the AnswerKey parameter of the ScoreMyKnownAnswers/2011-09-01 Review Policy.

PolicyParameter
Name of the parameter from the Review policy.

Qualification
The Qualification data structure represents a Qualification assigned to a user, including the Qualification type and the value (score).

QualificationRequest
The QualificationRequest data structure represents a request a Worker has made for a Qualification.

QualificationRequirement
The QualificationRequirement data structure describes a Qualification that a Worker must have before the Worker is allowed to accept a HIT. A requirement may optionally state that a Worker must have the Qualification in order to preview the HIT, or see the HIT in search results.

QualificationType
The QualificationType data structure represents a Qualification type, a description of a property of a Worker that must match the requirements of a HIT for the Worker to be able to accept the HIT. The type also describes how a Worker can obtain a Qualification of that type, such as through a Qualification test.

RejectAssignmentRequest

RejectAssignmentResponse

RejectQualificationRequestRequest

RejectQualificationRequestResponse

ReviewActionDetail
Both the AssignmentReviewReport and the HITReviewReport elements contains the ReviewActionDetail data structure. This structure is returned multiple times for each action specified in the Review Policy.

ReviewPolicy
HIT Review Policy data structures represent HIT review policies, which you specify when you create a HIT.

ReviewReport
Contains both ReviewResult and ReviewAction elements for a particular HIT.

ReviewResultDetail
This data structure is returned multiple times for each result specified in the Review Policy.

SendBonusRequest

SendBonusResponse

SendTestEventNotificationRequest

SendTestEventNotificationResponse

UpdateExpirationForHITRequest

UpdateExpirationForHITResponse

UpdateHITReviewStatusRequest

UpdateHITReviewStatusResponse

UpdateHITTypeOfHITRequest

UpdateHITTypeOfHITResponse

UpdateNotificationSettingsRequest

UpdateNotificationSettingsResponse

UpdateQualificationTypeRequest

UpdateQualificationTypeResponse

WorkerBlock
The WorkerBlock data structure represents a Worker who has been blocked. It has two elements: the WorkerId and the Reason for the block.

Enums

AcceptQualificationRequestError
Errors returned by AcceptQualificationRequest

ApproveAssignmentError
Errors returned by ApproveAssignment

AssociateQualificationWithWorkerError
Errors returned by AssociateQualificationWithWorker

CreateAdditionalAssignmentsForHITError
Errors returned by CreateAdditionalAssignmentsForHIT

CreateHITError
Errors returned by CreateHIT

CreateHITTypeError
Errors returned by CreateHITType

CreateHITWithHITTypeError
Errors returned by CreateHITWithHITType

CreateQualificationTypeError
Errors returned by CreateQualificationType

CreateWorkerBlockError
Errors returned by CreateWorkerBlock

DeleteHITError
Errors returned by DeleteHIT

DeleteQualificationTypeError
Errors returned by DeleteQualificationType

DeleteWorkerBlockError
Errors returned by DeleteWorkerBlock

DisassociateQualificationFromWorkerError
Errors returned by DisassociateQualificationFromWorker

GetAccountBalanceError
Errors returned by GetAccountBalance

GetAssignmentError
Errors returned by GetAssignment

GetFileUploadURLError
Errors returned by GetFileUploadURL

GetHITError
Errors returned by GetHIT

GetQualificationScoreError
Errors returned by GetQualificationScore

GetQualificationTypeError
Errors returned by GetQualificationType

ListAssignmentsForHITError
Errors returned by ListAssignmentsForHIT

ListBonusPaymentsError
Errors returned by ListBonusPayments

ListHITsError
Errors returned by ListHITs

ListHITsForQualificationTypeError
Errors returned by ListHITsForQualificationType

ListQualificationRequestsError
Errors returned by ListQualificationRequests

ListQualificationTypesError
Errors returned by ListQualificationTypes

ListReviewPolicyResultsForHITError
Errors returned by ListReviewPolicyResultsForHIT

ListReviewableHITsError
Errors returned by ListReviewableHITs

ListWorkerBlocksError
Errors returned by ListWorkerBlocks

ListWorkersWithQualificationTypeError
Errors returned by ListWorkersWithQualificationType

NotifyWorkersError
Errors returned by NotifyWorkers

RejectAssignmentError
Errors returned by RejectAssignment

RejectQualificationRequestError
Errors returned by RejectQualificationRequest

SendBonusError
Errors returned by SendBonus

SendTestEventNotificationError
Errors returned by SendTestEventNotification

UpdateExpirationForHITError
Errors returned by UpdateExpirationForHIT

UpdateHITReviewStatusError
Errors returned by UpdateHITReviewStatus

UpdateHITTypeOfHITError
Errors returned by UpdateHITTypeOfHIT

UpdateNotificationSettingsError
Errors returned by UpdateNotificationSettings

UpdateQualificationTypeError
Errors returned by UpdateQualificationType

Traits

MechanicalTurk
Trait representing the capabilities of the Amazon MTurk API. Amazon MTurk clients implement this trait.